Remember the current directory stack name in a file.
With this, when a new terminal is opened on a machine, it uses the directory
stack that was selected last in *any* of the other terminals.

Also, the script now checks if a directory stack file exists, before switching
to it. If it does not exist, it asks if the user wants to create it.